Fix #63 cannot exit chat screen
The third-party code and xib in JSQMessagesViewController have been modified, contrary to best practice, and it turns out that the back button was given incorrect placement that put it off-screen at runtime.

This has been corrected in situ for 1.0, but going forward this view should be reimplemented properly, that is without modifying the third-party library.