Add `Pods.framework` back to the `Nudj beta` target
Another solution would be to specify both targets explicitly in the Podfile, but that leads to needless duplication xcconfig files, etc, and doesn't scale as more targets are added.

A better solution would be to migrate away from Cocoapods to something less intrusive such as Carthage or plain Git submodules.