#82 Amend URL encoding of search parameters
Search parameters were not being properly percent-encoded, so an NSURL could not be created if the search parameters contained illegal characters.

Fixed at the call site but now we need to audit all API endpoints as per #19 to cleanly separate the path and parameter components of all API calls so that each can be encoded correctly. Otherwise search terms with a "?" in them will be incompatible with the current bad practice of hardcoding parameters in what is supposed to be the path component of the URL.